Moved reusable code to rcptt.resource plugins
