+++ core/contexts/org.eclipse.rcptt.ctx.resources/src/org/eclipse/rcptt/ctx/resources/WorkspaceContextMaker.java
@@ -10,41 +10,36 @@
  *******************************************************************************/
 package org.eclipse.rcptt.ctx.resources;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
 
-import org.eclipse.core.resources.IFile;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.resource.Resource;
-
 import org.eclipse.rcptt.core.ContextType;
 import org.eclipse.rcptt.core.DefaultContextMaker;
-import org.eclipse.rcptt.core.Q7Features;
 import org.eclipse.rcptt.core.model.IContext;
 import org.eclipse.rcptt.core.model.IQ7NamedElement;
 import org.eclipse.rcptt.core.model.ModelException;
 import org.eclipse.rcptt.core.persistence.IPersistenceModel;
 import org.eclipse.rcptt.core.scenario.Context;
-import org.eclipse.rcptt.ctx.internal.resources.Activator;
 import org.eclipse.rcptt.internal.core.RcpttPlugin;
 import org.eclipse.rcptt.resources.ImportUtils;
 import org.eclipse.rcptt.resources.WSUtils;
-import org.eclipse.rcptt.workspace.WSFile;
-import org.eclipse.rcptt.workspace.WSFileLink;
-import org.eclipse.rcptt.workspace.WSFolder;
-import org.eclipse.rcptt.workspace.WSFolderLink;
+import org.eclipse.rcptt.resources.WorkspaceDataImporter;
+import org.eclipse.rcptt.resources.WorkspaceDataMaker;
 import org.eclipse.rcptt.workspace.WSProject;
 import org.eclipse.rcptt.workspace.WSRoot;
 import org.eclipse.rcptt.workspace.WorkspaceContext;
+import org.eclipse.rcptt.workspace.WorkspaceData;
 import org.eclipse.rcptt.workspace.WorkspaceFactory;
 
 public class WorkspaceContextMaker extends DefaultContextMaker {
+
+	private WorkspaceDataMaker maker;
+
+	public WorkspaceContextMaker() {
+		this.maker = new WorkspaceDataMaker();
+	}
 
 	@Override
 	public Context create(Resource resource, ContextType type, String name) {
@@ -57,151 +52,15 @@ public class WorkspaceContextMaker extends DefaultContextMaker {
 	@Override
 	public void makeExecutable(Context context, IQ7NamedElement source)
 			throws ModelException {
-		makeSelfSufficientContext((WorkspaceContext) context,
+		maker.makeSelfSufficientData((WorkspaceData) context,
 				source.getModifiedPersistenceModel(), source);
-	}
-
-	public void makeSelfSufficientContext(WorkspaceContext context,
-			IPersistenceModel model, IQ7NamedElement source)
-			throws ModelException {
-		if (model != null) {
-			model.extractAll();
-		}
-		// Load and store all required resources into context object
-		WSRoot root = context.getContent();
-		if (root != null) {
-			for (WSProject wsProject : root.getProjects()) {
-				String location = source.getParent().getResource()
-						.getLocation().toString();
-				fillFilesContent(wsProject.getFiles(), location, model);
-				fillFileLinksContent(wsProject.getFileLinks());
-
-				EList<WSFolder> folders = wsProject.getFolders();
-				processFolders(folders, location, model);
-
-				fillFolderLinksContent(wsProject.getFolderLinks(), model);
-			}
-
-			fillFolderLinksContent(root.getProjectLinks(), model);
-		}
-	}
-
-	private void processFolders(EList<WSFolder> folders, String location,
-			IPersistenceModel model) throws ModelException {
-		for (WSFolder wsFolder : folders) {
-			EList<WSFolder> list = wsFolder.getFolders();
-			processFolders(list, location, model);
-
-			EList<WSFile> files = wsFolder.getFiles();
-			fillFilesContent(files, location, model);
-
-			EList<WSFileLink> fileLinks = wsFolder.getFileLinks();
-			fillFileLinksContent(fileLinks);
-
-			fillFolderLinksContent(wsFolder.getFolderLinks(), model);
-		}
-	}
-
-	private void fillFilesContent(EList<WSFile> files, String location,
-			IPersistenceModel model) {
-		for (WSFile wsFile : files) {
-			fillFile(location, wsFile, model);
-		}
-	}
-
-	private void fillFileLinksContent(EList<WSFileLink> fileLinks)
-			throws ModelException {
-		for (WSFileLink fileLink : fileLinks)
-			fillFileLink(fileLink);
-	}
-
-	private <T extends WSFolderLink> void fillFolderLinksContent(
-			EList<T> folderLinks, IPersistenceModel model)
-			throws ModelException {
-		for (WSFolderLink folderLink : folderLinks)
-			fillFolderLink(folderLink, model);
-	}
-
-	private void fillFolderLink(WSFolderLink folderLink, IPersistenceModel model)
-			throws ModelException {
-		try {
-			WSUtils.loadFolderLinkWithContent(folderLink);
-		} catch (IOException e) {
-			Activator.log(e);
-		} catch (CoreException e) {
-			Activator.log(e);
-		}
-	}
-
-	private void fillFile(String location, WSFile wsFile,
-			IPersistenceModel model) {
-		try {
-			final byte[] content = WSUtils.getFileContent(location, wsFile,
-					model);
-			if (Q7Features.getInstance().isTrue(
-					Q7Features.Q7_CONTEXTS_RESOURCES_ZIPPED_TRANSFER)) {
-
-				ByteArrayOutputStream bout = new ByteArrayOutputStream();
-				ZipOutputStream zout = new ZipOutputStream(bout);
-				zout.setLevel(9);
-				zout.setMethod(ZipOutputStream.DEFLATED);
-				ZipEntry entry = new ZipEntry("content");
-				entry.setTime(1);
-				zout.putNextEntry(entry);
-				zout.write(content);
-
-				zout.close();
-				wsFile.setContent(bout.toByteArray());
-			} else {
-				wsFile.setContent(content);
-			}
-		} catch (IOException e) {
-			Activator.log(e);
-		}
-	}
-
-	private void fillFileLink(WSFileLink fileLink) throws ModelException {
-		WSUtils.loadFileLinkWithContent(fileLink);
-	}
-
-	public void makeExecutableContext(WSFile child, IFile iFile) {
-		if ("false".equals(Q7Features.getInstance().getValue(
-				Q7Features.Q7_CONTEXTS_RESOURCES_TRANSFER_CONTENT))) {
-			return;
-		}
-		try {
-			InputStream contents = iFile.getContents();
-			final byte[] content = WSUtils.getStreamContent(contents);
-			contents.close();
-			if (Q7Features.getInstance().isTrue(
-					Q7Features.Q7_CONTEXTS_RESOURCES_ZIPPED_TRANSFER)) {
-
-				ByteArrayOutputStream bout = new ByteArrayOutputStream();
-				ZipOutputStream zout = new ZipOutputStream(bout);
-				zout.setLevel(9);
-				zout.setMethod(ZipOutputStream.DEFLATED);
-				ZipEntry entry = new ZipEntry("content");
-				entry.setTime(1);
-				zout.putNextEntry(entry);
-				zout.write(content);
-				zout.close();
-				child.setContent(bout.toByteArray());
-			} else {
-				child.setContent(content);
-			}
-		} catch (IOException e) {
-			Activator.log(e);
-		} catch (CoreException e) {
-			Activator.log(e);
-		}
 	}
 
 	@Override
 	public void captureContents(IContext context, Context data,
 			IProgressMonitor monitor) throws CoreException {
 		WorkspaceContext wsContext = (WorkspaceContext) data;
-		WorkspaceContextImporter importer = new WorkspaceContextImporter(
-				wsContext);
+		WorkspaceDataImporter importer = new WorkspaceDataImporter(wsContext);
 
 		IPersistenceModel model = context.getPersistenceModel();
 		model.removeAll();
