+++ core/verifications/org.eclipse.rcptt.verifications.resources/src/org/eclipse/rcptt/verifications/resources/WorkspaceVerificationMaker.java
@@ -10,20 +10,12 @@
  *******************************************************************************/
 package org.eclipse.rcptt.verifications.resources;
 
-import java.io.ByteArrayOutputStream;
-import java.io.IOException;
-import java.io.InputStream;
 import java.util.List;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipOutputStream;
 
-import org.eclipse.core.resources.IFile;
 import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.rcptt.core.DefaultVerificationMaker;
-import org.eclipse.rcptt.core.Q7Features;
 import org.eclipse.rcptt.core.VerificationType;
 import org.eclipse.rcptt.core.model.IQ7NamedElement;
 import org.eclipse.rcptt.core.model.IVerification;
@@ -33,17 +25,20 @@ import org.eclipse.rcptt.core.scenario.Verification;
 import org.eclipse.rcptt.internal.core.RcpttPlugin;
 import org.eclipse.rcptt.resources.ImportUtils;
 import org.eclipse.rcptt.resources.WSUtils;
-import org.eclipse.rcptt.verifications.internal.resources.Activator;
-import org.eclipse.rcptt.workspace.WSFile;
-import org.eclipse.rcptt.workspace.WSFileLink;
-import org.eclipse.rcptt.workspace.WSFolder;
-import org.eclipse.rcptt.workspace.WSFolderLink;
+import org.eclipse.rcptt.resources.WorkspaceDataImporter;
+import org.eclipse.rcptt.resources.WorkspaceDataMaker;
 import org.eclipse.rcptt.workspace.WSProject;
 import org.eclipse.rcptt.workspace.WSRoot;
 import org.eclipse.rcptt.workspace.WorkspaceFactory;
 import org.eclipse.rcptt.workspace.WorkspaceVerification;
 
 public class WorkspaceVerificationMaker extends DefaultVerificationMaker {
+
+	private WorkspaceDataMaker maker;
+
+	public WorkspaceVerificationMaker() {
+		this.maker = new WorkspaceDataMaker();
+	}
 
 	@Override
 	public Verification create(Resource resource, VerificationType type, String name) {
@@ -56,147 +51,15 @@ public class WorkspaceVerificationMaker extends DefaultVerificationMaker {
 	@Override
 	public void makeExecutable(Verification verification, IQ7NamedElement source)
 			throws ModelException {
-		makeSelfSufficientVerification((WorkspaceVerification) verification,
+		maker.makeSelfSufficientData((WorkspaceVerification) verification,
 				source.getModifiedPersistenceModel(), source);
-	}
-
-	public void makeExecutableVerification(WSFile child, IFile iFile) {
-		try {
-			InputStream contents = iFile.getContents();
-			final byte[] content = WSUtils.getStreamContent(contents);
-			contents.close();
-			if (Q7Features.getInstance().isTrue(
-					Q7Features.Q7_CONTEXTS_RESOURCES_ZIPPED_TRANSFER)) {
-
-				ByteArrayOutputStream bout = new ByteArrayOutputStream();
-				ZipOutputStream zout = new ZipOutputStream(bout);
-				zout.setLevel(9);
-				zout.setMethod(ZipOutputStream.DEFLATED);
-				ZipEntry entry = new ZipEntry("content");
-				entry.setTime(1);
-				zout.putNextEntry(entry);
-				zout.write(content);
-				zout.close();
-				child.setContent(bout.toByteArray());
-			} else {
-				child.setContent(content);
-			}
-		} catch (IOException e) {
-			Activator.log(e);
-		} catch (CoreException e) {
-			Activator.log(e);
-		}
-	}
-
-	public void makeSelfSufficientVerification(WorkspaceVerification verification,
-			IPersistenceModel model, IQ7NamedElement source)
-			throws ModelException {
-		if (model != null) {
-			model.extractAll();
-		}
-		// Load and store all required resources into context object
-		WSRoot root = verification.getContent();
-		if (root != null) {
-			for (WSProject wsProject : root.getProjects()) {
-				String location = source.getParent().getResource()
-						.getLocation().toString();
-				fillFilesContent(wsProject.getFiles(), location, model);
-				fillFileLinksContent(wsProject.getFileLinks());
-
-				EList<WSFolder> folders = wsProject.getFolders();
-				processFolders(folders, location, model);
-
-				fillFolderLinksContent(wsProject.getFolderLinks(), model);
-			}
-
-			fillFolderLinksContent(root.getProjectLinks(), model);
-		}
-	}
-
-	private void processFolders(EList<WSFolder> folders, String location,
-			IPersistenceModel model) throws ModelException {
-		for (WSFolder wsFolder : folders) {
-			EList<WSFolder> list = wsFolder.getFolders();
-			processFolders(list, location, model);
-
-			EList<WSFile> files = wsFolder.getFiles();
-			fillFilesContent(files, location, model);
-
-			EList<WSFileLink> fileLinks = wsFolder.getFileLinks();
-			fillFileLinksContent(fileLinks);
-
-			fillFolderLinksContent(wsFolder.getFolderLinks(), model);
-		}
-	}
-
-	private void fillFilesContent(EList<WSFile> files, String location,
-			IPersistenceModel model) {
-		for (WSFile wsFile : files) {
-			fillFile(location, wsFile, model);
-		}
-	}
-
-	private void fillFileLinksContent(EList<WSFileLink> fileLinks)
-			throws ModelException {
-		for (WSFileLink fileLink : fileLinks)
-			fillFileLink(fileLink);
-	}
-
-	private <T extends WSFolderLink> void fillFolderLinksContent(
-			EList<T> folderLinks, IPersistenceModel model)
-			throws ModelException {
-		for (WSFolderLink folderLink : folderLinks)
-			fillFolderLink(folderLink, model);
-	}
-
-	private void fillFolderLink(WSFolderLink folderLink, IPersistenceModel model)
-			throws ModelException {
-		try {
-			WSUtils.loadFolderLinkWithContent(folderLink);
-		} catch (IOException e) {
-			Activator.log(e);
-		} catch (CoreException e) {
-			Activator.log(e);
-		}
-	}
-
-	private void fillFile(String location, WSFile wsFile,
-			IPersistenceModel model) {
-		try {
-			final byte[] content = WSUtils.getFileContent(location, wsFile,
-					model);
-			if (Q7Features.getInstance().isTrue(
-					Q7Features.Q7_CONTEXTS_RESOURCES_ZIPPED_TRANSFER)) {
-
-				ByteArrayOutputStream bout = new ByteArrayOutputStream();
-				ZipOutputStream zout = new ZipOutputStream(bout);
-				zout.setLevel(9);
-				zout.setMethod(ZipOutputStream.DEFLATED);
-				ZipEntry entry = new ZipEntry("content");
-				entry.setTime(1);
-				zout.putNextEntry(entry);
-				zout.write(content);
-
-				zout.close();
-				wsFile.setContent(bout.toByteArray());
-			} else {
-				wsFile.setContent(content);
-			}
-		} catch (IOException e) {
-			Activator.log(e);
-		}
-	}
-
-	private void fillFileLink(WSFileLink fileLink) throws ModelException {
-		WSUtils.loadFileLinkWithContent(fileLink);
 	}
 
 	@Override
 	public void captureContents(IVerification verifiacion, Verification data,
 			IProgressMonitor monitor) throws CoreException {
 		WorkspaceVerification verification = (WorkspaceVerification) data;
-		WorkspaceVerificationImporter importer = new WorkspaceVerificationImporter(
-				verification);
+		WorkspaceDataImporter importer = new WorkspaceDataImporter(verification);
 
 		IPersistenceModel model = verifiacion.getPersistenceModel();
 		model.removeAll();
