+++ rcp/org.eclipse.rcptt.resources.ui/META-INF/MANIFEST.MF
@@ -9,6 +9,7 @@ Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Bundle-ActivationPolicy: lazy
 Export-Package: org.eclipse.rcptt.resources.ui.actions,
  org.eclipse.rcptt.resources.ui.imports,
+ org.eclipse.rcptt.resources.ui.viewers,
  org.eclipse.rcptt.resources.ui.wizards
 Require-Bundle: org.eclipse.jface,
  org.eclipse.ui.ide,
