+++ rcp/org.eclipse.rcptt.resources.ui/src/org/eclipse/rcptt/resources/ui/viewers/WorkspaceLabelProvider.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     Xored Software Inc - initial API and implementation and/or initial documentation
  *******************************************************************************/
-package org.eclipse.rcptt.verifications.resources.ui.viewers;
+package org.eclipse.rcptt.resources.ui.viewers;
 
 import org.eclipse.jface.viewers.BaseLabelProvider;
 import org.eclipse.jface.viewers.DelegatingStyledCellLabelProvider.IStyledLabelProvider;
