+++ runtime/org.eclipse.rcptt.resources.impl/META-INF/MANIFEST.MF
@@ -10,6 +10,7 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.core.filebuffers;resolution:=optional,
  org.eclipse.core.filesystem;resolution:=optional,
  org.eclipse.rcptt.core;bundle-version="[2.3.0,3.0.0)",
+ org.eclipse.rcptt.resources;bundle-version="[2.3.0,3.0.0)",
  org.eclipse.rcptt.tesla.ecl.impl;bundle-version="[2.3.0,3.0.0)"
 Bundle-Activator: org.eclipse.rcptt.resources.internal.impl.Activator
 Export-Package: org.eclipse.rcptt.resources.impl
